Threshold getter, refactoring structure
