+++ multisig-wallet/project/multisig-contract/tests/functions/core/mod.rs
@@ -1,0 +1,4 @@
+mod cancel_transaction;
+mod constructor;
+mod execute_transaction;
+mod transfer;
