+++ multisig-wallet/project/multisig-contract/tests/functions/info/mod.rs
@@ -1,0 +1,4 @@
+mod balance;
+mod nonce;
+mod threshold;
+mod transaction_hash;
