+++ multisig-wallet/project/multisig-contract/tests/functions/info/threshold.rs
@@ -1,0 +1,18 @@
+mod success {
+
+    use crate::utils::{
+        interface::{core::constructor, info::threshold},
+        setup::{default_users, setup_env, DEFAULT_THRESHOLD, VALID_SIGNER_PK},
+    };
+
+    #[tokio::test]
+    async fn returns_threshold() {
+        let (_private_key, deployer, _non_owner) = setup_env(VALID_SIGNER_PK).await.unwrap();
+
+        constructor(&deployer.contract, default_users(), DEFAULT_THRESHOLD).await;
+
+        let current_threshold = threshold(&deployer.contract).await.value;
+
+        assert_eq!(DEFAULT_THRESHOLD, current_threshold);
+    }
+}
