+++ multisig-wallet/project/multisig-contract/tests/functions/mod.rs
@@ -1,7 +1,2 @@
-mod balance;
-mod cancel_transaction;
-mod constructor;
-mod execute_transaction;
-mod nonce;
-mod transaction_hash;
-mod transfer;
+mod core;
+mod info;
